backend/requirement.txt
absl-py==2.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.7
aiohttp-retry==2.9.1
aiosignal==1.3.2
alembic==1.15.1
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.5
attrs==24.3.0
audioread==3.0.1
babel==2.17.0
backports.tarfile==1.2.0
bcrypt==4.3.0
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cryptography==44.0.3
cycler==0.12.1
debugpy==1.8.14
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
dnspython==2.8.0
docopt==0.6.2
docutils==0.21.2
easyocr==1.7.2
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.17.0
fire==0.7.0
flasgger==0.9.7.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.7.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-MySQL==1.6.0
Flask-MySQLdb==2.0.0
Flask-PyMongo==3.0.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
flatbuffers==25.1.24
fonttools==4.55.3
fqdn==1.5.1
frozenlist==1.6.2
fsspec==2025.2.0
gast==0.6.0
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.70.0
gunicorn==23.0.0
h11==0.16.0
h5py==3.12.1
httpcore==1.0.9
httpx==0.28.1
id==1.5.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.6.1
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.10.0
joblib==1.4.2
Js2Py==0.74
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
keras==3.8.0
keyring==25.6.0
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
librosa==0.8.1
llvmlite==0.44.0
Mako==1.3.9
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplot==0.1.9
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
ml-dtypes==0.4.1
more-itertools==10.6.0
moviepy==2.2.1
mpmath==1.3.0
multidict==6.4.4
mysql-connector-python==9.1.0
mysqlclient==2.2.6
namex==0.0.8
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nh3==0.2.21
ninja==1.11.1.3
nltk==3.9.1
notebook==7.4.2
notebook_shim==0.2.4
numba==0.61.0
numpy==2.0.2
ollama==0.5.1
openai==0.28.0
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt_einsum==3.4.0
optree==0.14.0
outcome==1.3.0.post0
overrides==7.7.0
p2j==1.3.2
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pillow==10.4.0
pipwin==0.5.2
platformdirs==4.3.6
pooch==1.8.2
proglog==0.1.12
prometheus_client==0.21.1
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==3.20.3
psutil==7.0.0
pure_eval==0.2.3
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.1
pyjsparser==2.7.1
PyJWT==2.10.1
pyloco==0.0.139
pymongo==4.8.0
PyMySQL==1.1.1
pyparsing==3.2.1
PyPrind==2.11.3
pySmartDL==1.3.4
PySocks==1.7.1
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.3.0
python-whois==0.9.5
pytz==2024.2
pywin32==310
pywin32-ctypes==0.2.3
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==26.4.0
readme_renderer==44.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
resampy==0.4.3
rfc3339-validator==0.1.4
rfc3986==2.0.0
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.24.0
scikit-image==0.25.1
scikit-learn==1.6.0
scipy==1.15.0
selenium==4.27.1
Send2Trash==1.8.3
shapely==2.0.7
SimpleWebSocketServer==0.1.2
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
soupsieve==2.6
SQLAlchemy==2.0.36
stack-data==0.6.3
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==2.5.0
terminado==0.18.1
textblob==0.19.0
threadpoolctl==3.5.0
tifffile==2025.1.10
tinycss2==1.4.0
torch==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
trio==0.27.0
trio-websocket==0.11.1
twilio==9.6.2
twine==6.1.0
types-python-dateutil==2.9.0.20241206
typing==3.7.4.3
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
uri-template==1.3.0
urllib3==2.2.3
ushlex==0.99.1
vaderSentiment==3.3.2
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
WTForms==3.2.1
XlsxWriter==3.2.0
yarl==1.20.0
zipp==3.21.0
